:art: [#1218] -- Clean up tests/code for ZaakEigenschap write operations
