+++ src/openzaak/components/zaken/tests/test_notifications_send.py
@@ -151,9 +151,7 @@ class SendNotifTestCase(NotificationServiceMixin, JWTAuthMixin, APITestCase):
         """
 
         client = mock_client.return_value
-        zaaktype = ZaakTypeFactory.create(concept=False)
-        zaaktype_url = reverse(zaaktype)
-        zaak = ZaakFactory.create(zaaktype=f"http://testserver{zaaktype_url}")
+        zaak = ZaakFactory.create()
         zaak_url = get_operation_url("zaak_read", uuid=zaak.uuid)
         zaakobject = ZaakObjectFactory.create(zaak=zaak, relatieomschrijving="old")
         zaakobject_url = get_operation_url("zaakobject_update", uuid=zaakobject.uuid)
@@ -173,7 +171,7 @@ class SendNotifTestCase(NotificationServiceMixin, JWTAuthMixin, APITestCase):
                 "aanmaakdatum": "2012-01-14T00:00:00Z",
                 "kenmerken": {
                     "bronorganisatie": zaak.bronorganisatie,
-                    "zaaktype": f"http://testserver{zaaktype_url}",
+                    "zaaktype": f"http://testserver{reverse(zaak.zaaktype)}",
                     "vertrouwelijkheidaanduiding": zaak.vertrouwelijkheidaanduiding,
                 },
             },
@@ -185,9 +183,7 @@ class SendNotifTestCase(NotificationServiceMixin, JWTAuthMixin, APITestCase):
         Check if notifications will be send when zaak-eigenschap is updated
         """
         client = mock_client.return_value
-        zaaktype = ZaakTypeFactory.create(concept=False)
-        zaaktype_url = reverse(zaaktype)
-        zaak = ZaakFactory.create(zaaktype=f"http://testserver{zaaktype_url}")
+        zaak = ZaakFactory.create()
         zaak_url = get_operation_url("zaak_read", uuid=zaak.uuid)
         zaakeigenschap = ZaakEigenschapFactory.create(zaak=zaak, waarde="old")
         zaakeigenschap_url = get_operation_url(
@@ -209,7 +205,7 @@ class SendNotifTestCase(NotificationServiceMixin, JWTAuthMixin, APITestCase):
                 "aanmaakdatum": "2012-01-14T00:00:00Z",
                 "kenmerken": {
                     "bronorganisatie": zaak.bronorganisatie,
-                    "zaaktype": f"http://testserver{zaaktype_url}",
+                    "zaaktype": f"http://testserver{reverse(zaak.zaaktype)}",
                     "vertrouwelijkheidaanduiding": zaak.vertrouwelijkheidaanduiding,
                 },
             },
